Break up ValueNode.cs so each Value has it's own file
