+++ src/linker/Linker.Dataflow/FieldValue.cs
@@ -1,0 +1,41 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using ILLink.Shared.DataFlow;
+using Mono.Linker;
+using Mono.Linker.Dataflow;
+using FieldDefinition = Mono.Cecil.FieldDefinition;
+using TypeDefinition = Mono.Cecil.TypeDefinition;
+
+
+namespace ILLink.Shared.TrimAnalysis
+{
+
+	/// <summary>
+	/// A representation of a field. Typically a result of ldfld.
+	/// </summary>
+	sealed partial record FieldValue : IValueWithStaticType
+	{
+		public FieldValue (TypeDefinition? staticType, FieldDefinition fieldToLoad, DynamicallyAccessedMemberTypes dynamicallyAccessedMemberTypes)
+		{
+			StaticType = staticType;
+			Field = fieldToLoad;
+			DynamicallyAccessedMemberTypes = dynamicallyAccessedMemberTypes;
+		}
+
+		public readonly FieldDefinition Field;
+
+		public override DynamicallyAccessedMemberTypes DynamicallyAccessedMemberTypes { get; }
+
+		public override IEnumerable<string> GetDiagnosticArgumentsForAnnotationMismatch ()
+			=> new string[] { Field.GetDisplayName () };
+
+		public TypeDefinition? StaticType { get; }
+
+		public override SingleValue DeepCopy () => this; // This value is immutable
+
+		public override string ToString () => this.ValueToString (Field, DynamicallyAccessedMemberTypes);
+	}
+}
